Fixed hibernate bugs

+++ src/main/java/com/kisanmitra/dto/OrderItemRequest.java
@@ -3,12 +3,14 @@ package com.kisanmitra.dto;
 import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
+import java.math.BigDecimal;
+
 
 @Data
 @Getter
 @Setter
 public class OrderItemRequest {
     private Long listingId;
-    private Double quantity;
+    private BigDecimal quantity;
 }
